+++ src/javascript/constructors/contextmenu.js
@@ -1,5 +1,6 @@
 import { puffin } from '@mkenzo_8/puffin'
 import RunningConfig from 'RunningConfig'
+import ContextMenuWrapper from '../components/context.menu'
 
 function ContextMenu({
 	list,
@@ -8,40 +9,6 @@ function ContextMenu({
 	x,
 	y
 }){
-	const ContextWrapper = puffin.style.div`
-		&{
-			background:var(--contextmenuBackground);
-			padding:5px;
-			position:fixed;
-			color:white;
-			border-radius:5px;
-			box-shadow:0px 0px 3px rgba(0,0,0,0.2);
-			display:block;
-		}
-		& > button{
-			background:var(--contextmenuButtonBackground);
-			color:var(--contextmenuButtonText);
-			border:0;
-			padding:6px;
-			outline:0;
-			border-radius:5px;
-			display:block;
-			width:100%;
-			text-align:left;
-		}
-		& >  button:hover{
-			background:var(--contextmenuButtonHoveringBackground);
-			color:var(--contextmenuButtonHoveringText);
-		}
-		& >  span{
-			height:1.5px;
-			border-radius:25px;
-			width:95%;
-			display:block;
-			background:var(--contextmenuDivider);
-			margin:3px auto;
-		}
-	`;
 	const randomSelector = Math.random()
 	const contextMenusCreated = document.getElementsByClassName("contextMenu");
 	const computedMethods = {...list.map(a=>a.action)}
@@ -53,7 +20,7 @@ function ContextMenu({
 		y:event!= null?event.pageY:y,
 	}
 	const ContextComponent = puffin.element(`
-		<ContextWrapper id="${randomSelector}" class="contextMenu" style="top:${positions.y}px; left:${positions.x};">
+		<ContextMenuWrapper id="${randomSelector}" class="contextMenu" style="top:${positions.y}px; left:${positions.x};">
 		${(function(){
 				let content = "";
 				list.map((option,index)=>{
@@ -65,11 +32,11 @@ function ContextMenu({
 				})
 				return content
 			})()}
-		</ContextWrapper>
+		</ContextMenuWrapper>
 	`,{
 		methods:computedMethods,
 		components:{
-			ContextWrapper
+			ContextMenuWrapper
 		},
 		events:{
 			mounted(target){