+++ src/javascript/constructors/menu.js
@@ -1,6 +1,7 @@
 import { puffin } from '@mkenzo_8/puffin'
 import Dropmenu  from '../components/dropmenu'
 import requirePath from '../utils/require'
+import ArrowIcon from '../components/icons/arrow'
 
 const { remote } = requirePath("electron")
 
@@ -11,35 +12,82 @@ let MenuBarMacOS;
 	MenuBarMacOS  = new Menu()
 })()
 
+function closeAllSubmenus(parent){
+	const subMenusOpened = Object.keys(parent.getElementsByClassName("submenu")).map((ele)=>{
+		return parent.getElementsByClassName("submenu")[ele]
+	})
+	subMenusOpened.map((element)=>{
+		element.remove()
+	})
+}
+
+function getDropmenu(list){
+	return `
+	<div>
+	${(function(){
+		let content = "";
+		list.map(function(option,index){
+			if(option.label !== undefined){
+				content += `
+					<div>
+						<a 
+							${option.hint != null?`title="${option.hint}"`:""} 
+							click="${!option.list?`$${index}`:''}" 
+							mouseenter="${option.list?`$${index}`:'$hideMenus'}">
+							${option.label}
+							${option.list?'<ArrowIcon/>':''}
+						</a>
+					</div>`
+			}else{
+				content += `<span/>`
+			}
+		})
+		return content;
+	})()}
+	</div>`
+}
+
+function getMenu(button,list,leftMargin){
+	const methodsToBind = Object.assign({},list.map((option)=> {
+		if(option.action){
+			return option.action
+		}else if(option.list){
+			return function(e){
+				closeAllSubmenus(e.target.parentElement)
+				const subMenuComponent = getMenu(
+					null,
+					option.list,
+					e.target.clientWidth+10
+				)
+				puffin.render(subMenuComponent,e.target.parentElement)
+			}
+		}
+	}))
+	return puffin.element(`
+		<Dropmenu class="${button?'':'submenu'}" submenu="${button?false:true}" style="${button?'':`position:absolute;margin-top:-10px;margin-left:${leftMargin}px;`}">
+			${button?`<button mouseover="$hideMenus" click="$hideMenus">${button}</button>`:''}
+			${getDropmenu(list)}
+		</Dropmenu>
+	`,{
+		components:{
+			Dropmenu,
+			ArrowIcon
+		},
+		methods:{
+			...methodsToBind,
+			hideMenus(){
+				closeAllSubmenus(this.parentElement.parentElement)
+			}
+		}
+	})
+}
+
 function Menu({
 	button,
 	list
 }){
 	if(eval('process.platform') !== "darwin"){
-		const methodsToBind = Object.assign({},list.map((option)=> option.action))
-		const MenuComponent = puffin.element(`
-			<Dropmenu>
-				<button>${button}</button>
-				<div>
-					${(function(){
-						let content = "";
-						list.map(function(option,index){
-							if(option.label !== undefined){
-								content += `<a ${option.hint != null?`title="${option.hint}"`:""} click="$${index}">${option.label}</a>`
-							}else{
-								content += `<span/>`
-							}
-						})
-						return content;
-					})()}
-				</div>
-			</Dropmenu>
-		`,{
-			components:{
-				Dropmenu
-			},
-			methods:methodsToBind
-		})
+		const MenuComponent = getMenu(button,list)
 		puffin.render(MenuComponent,document.getElementById("dropmenus"))
 	}else{
 		appendToBar(createTemplate(button,list))
@@ -48,23 +96,29 @@ function Menu({
 
 function createTemplate(button,list){
 	const { MenuItem } = remote
-
 	return new MenuItem({
 		label:button,
-		submenu:(function(){
-			return list.map((btn)=>{
-				if(btn.label !== undefined){
-					return {
-						label:btn.label,
-						click:btn.action
-					} 
-				}else{
-					return {
-						type: 'separator'
-					}
-				} 
-			})
-		})()
+		submenu:parseMenu(list)
+	})
+}
+
+function parseMenu(list){
+	return list.map((btn)=>{
+		if(btn.label && btn.action){
+			return {
+				label:btn.label,
+				click:btn.action
+			} 
+		} else if(btn.label && btn.list){
+			return {
+				label:btn.label,
+				submenu:parseMenu(btn.list)
+			} 
+		} else{
+			return {
+				type: 'separator'
+			}
+		} 
 	})
 }